requirements.txt
# Core
streamlit>=1.28.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
fpdf2>=2.7.0

# LangChain Core
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0

# Hugging Face Integration (Updated)
langchain-huggingface>=1.2.0
transformers>=4.35.0
sentence-transformers>=2.2.0
torch>=2.0.0
huggingface-hub>=0.20.0

# Vector Database
langchain-chroma>=0.1.0
chromadb>=0.4.0